Updated calculation of SMD
Updated calculation of SMD to report a value when only one of (but not both) comparator and referent groups lack covariate variation. In this case, the pooled SD is not used, rather the group with variation provides the SD value. Previously, the default behavior was to set SMD to missing when either group lacked covariate variation.